requirements.txt
accelerate==1.10.1
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
audioread==3.0.1
beautifulsoup4==4.13.5
braceexpand==0.1.7
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
decorator==5.2.1
docker-pycreds==0.4.0
exceptiongroup==1.3.0
fastapi==0.115.12
filelock==3.18.0
fsspec==2025.3.2
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0rc1
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
h5py==3.13.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.31.1
idna==3.10
Jinja2==3.1.6
joblib==1.5.0
laion_clap @ git+https://github.com/LAION-AI/CLAP.git@cc8f7654fc8b718434cf9ac6e6faf72f78c2797b
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.43.0
lyricsgenius==3.7.2
MarkupSafe==3.0.2
mpmath==1.3.0
msgpack==1.1.0
mutagen==1.47.0
networkx==3.4.2
numba==0.60.0
numpy==1.23.5
packaging==25.0
pandas==2.2.3
pillow==11.2.1
platformdirs==4.3.8
pooch==1.8.2
progressbar==2.5
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
regex==2024.11.6
requests==2.32.3
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentry-sdk==2.27.0
setproctitle==1.3.6
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.8
soxr==0.5.0.post1
spotipy==2.25.1
starlette==0.46.2
sympy==1.13.3
threadpoolctl==3.6.0
tiktoken==0.11.0
tokenizers==0.21.1
torch==2.8.0+cpu
torchaudio==2.8.0+cpu
torchlibrosa==0.1.0
torchvision==0.23.0+cpu
tqdm==4.67.1
transformers==4.52.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.3
wandb==0.19.11
watchfiles==1.1.0
wcwidth==0.2.13
webdataset==0.2.111
websockets==15.0.1
wget==3.2
